+++ src/utils/fullRoutes.ts
@@ -1,6 +1,7 @@
-import { FRONT, ADD_TASK, EDIT_TASK } from "./routeSegments";
+import { FRONT, ADD_TASK, EDIT_TASK, MANAGE_CATEGORIES } from "./routeSegments";
 
 // All valid routes within the application
 export const HOME_ROUTE = `/${FRONT}`;
 export const ADD_TASK_ROUTE = `/${FRONT}/${ADD_TASK}`;
 export const EDIT_TASK_ROUTE = `/${FRONT}/${EDIT_TASK}`;
+export const MANAGE_CATEGORIES_ROUTE = `/${FRONT}/${MANAGE_CATEGORIES}`;